+++ R/tm_t_summary.R
@@ -413,6 +413,7 @@ srv_summary <- function(id,
   shiny::moduleServer(id, function(input, output, session) {
     teal.code::init_chunks()
 
+    #validate(need())
     anl_selectors <- teal.transform::data_extract_multiple_srv(
       list(arm_var = arm_var, summarize_vars = summarize_vars),
       datasets = datasets
@@ -457,8 +458,16 @@ srv_summary <- function(id,
       input_summarize_vars <- anl_merged()$columns_source$summarize_vars
 
       shiny::validate(
+        shiny::need(
+          length(unique(anl_m$data()$USUBJID)) == nrow(anl_m$data()),
+          "Please choose an analysis dataset with unique USUBJID identifiers only"
+        ),
         shiny::need(input_arm_var, "Please select a treatment variable"),
         shiny::need(input_summarize_vars, "Please select a summarize variable"),
+        shiny::need(
+          !any(sapply(anl_m$data()[,input_summarize_vars], function(x) inherits(x, c("Date", "POSIXt")))),
+          "Date and POSIXt variables are not supported, please select other variables"
+        ),
         shiny::need(length(input_arm_var) <= 2, "Please limit treatment variables within two"),
         if (length(input_arm_var) == 2) {
           shiny::need(